Replace `StringsLoader` module with internal package in AnyLoader
* Move every class of Loader in a package called `internal`. None
of those classes should be accessible outside of the library.
* Move StringsLoader classes to the Loader module & delete it.
