+++ app/build.gradle
@@ -28,7 +28,7 @@ android {
 }
 
 dependencies {
-    implementation project(':stringsloader')
+    implementation project(':loader')
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation"org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
