+++ loader/src/main/java/com/romanwuattier/loader/ConverterType.kt
@@ -1,0 +1,5 @@
+package com.romanwuattier.loader
+
+enum class ConverterType {
+    JSON, OTHER
+}
