--- loader/src/main/java/com/romanwuattier/loader/converter/ConverterType.kt
@@ -1,5 +1,0 @@
-package com.romanwuattier.loader.converter
-
-enum class ConverterType {
-    JSON, OTHER
-}
