--- loader/src/main/java/com/romanwuattier/loader/data/LoadRequest.kt
@@ -1,25 +1,0 @@
-package com.romanwuattier.loader.data
-
-import android.content.Context
-import com.romanwuattier.loader.converter.ConverterType
-import java.io.File
-
-internal interface Request
-
-/**
- * A [RemoteRequest] consists in:
- * <ul>
- *     <li>Loading data from a specific remote [url]</li>
- *     <li>Deserializing data in a [ConcurrentHashMap] according to a [ConverterType]</li>
- * </ul>
- */
-internal data class RemoteRequest(val url: String, val cacheDir: File, val converterType: ConverterType) : Request
-
-/**
- * A [LocalRequest] consists in:
- * <ul>
- *     <li>Loading data from a specific file [path]</li>
- *     <li>Deserializing data in a [ConcurrentHashMap] according to a [ConverterType]</li>
- * </ul>
- */
-internal data class LocalRequest(val context: Context, val path: String, val converterType: ConverterType) : Request
