+++ loader/src/main/java/com/romanwuattier/loader/internal/converter/OtherConverter.kt
@@ -1,0 +1,9 @@
+package com.romanwuattier.loader.internal.converter
+
+import java.io.Reader
+import java.util.concurrent.ConcurrentHashMap
+
+internal class OtherConverter : ConverterStrategy {
+
+    override fun <K, V> convert(reader: Reader?): ConcurrentHashMap<K, V> = ConcurrentHashMap()
+}
