+++ loader/src/main/java/com/romanwuattier/loader/internal/retry/RetryStrategy.kt
@@ -1,0 +1,5 @@
+package com.romanwuattier.loader.internal.retry
+
+interface RetryStrategy {
+
+}
