+++ loader/src/main/java/com/romanwuattier/loader/internal/task/SaveTask.kt
@@ -1,0 +1,17 @@
+package com.romanwuattier.loader.internal.task
+
+import com.romanwuattier.loader.internal.store.Store
+import java.util.concurrent.Callable
+import java.util.concurrent.ConcurrentHashMap
+
+/**
+ * The [Callable] task to save remote data in the [Store.Memory]
+ */
+internal class SaveTask<K, V>(private val map: ConcurrentHashMap<K, V>,
+    private val memoryStore: Store.Memory<K, V>) : Callable<Boolean> {
+
+    override fun call(): Boolean {
+        memoryStore.putAll(map)
+        return true
+    }
+}
