--- loader/src/main/java/com/romanwuattier/loader/retry/RetryStrategy.kt
@@ -1,5 +1,0 @@
-package com.romanwuattier.loader.retry
-
-interface RetryStrategy {
-
-}
