--- loader/src/main/java/com/romanwuattier/loader/utils/Extensions.kt
@@ -1,9 +1,0 @@
-package com.romanwuattier.loader.utils
-
-import android.os.Looper
-
-fun checkMainThread() {
-    if (Thread.currentThread() != Looper.getMainLooper().thread) {
-        throw IllegalStateException("Execution was not running on the main thread.")
-    }
-}
